refactor(ui): phase B — domain API imports (#343)
Pages/hooks import from api/* modules; keep getApiErrorMessage and
resolveImageUrl on api barrel; fix tests mocks; CHANGELOG.

Made-with: Cursor
