+++ app/ui/src/App.routes.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import App from './App';
 
-vi.mock('./api/api', () => ({
+vi.mock('./api/systemAuditMetrics', () => ({
   trackSiteVisitor: vi.fn().mockResolvedValue(undefined),
 }));
