+++ app/ui/src/pages/FoodManagement/index.tsx
@@ -12,7 +12,8 @@ import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Checkbox from '@mui/material/Checkbox';
 import Info from '@mui/icons-material/Info';
-import { resolveImageUrl, fetchBirdFood, toggleBirdFood } from '../../api/api';
+import { resolveImageUrl } from '../../api/api';
+import { fetchBirdFood, toggleBirdFood } from '../../api/birdFoodFeed';
 import { queryKeys } from '../../api/queryKeys';
 import { BirdFood } from '../../types';
 import { PageHelp } from '../../components/PageHelp';
