+++ app/ui/src/pages/Library/DatasetExportsCard.tsx
@@ -16,7 +16,8 @@ import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import dayjs, { Dayjs } from 'dayjs';
-import { BASE_API_URL, exportDataset, retroExportDataset } from '../../api/api';
+import { BASE_API_URL } from '../../api/client';
+import { exportDataset, retroExportDataset } from '../../api/dataset';
 import { queryKeys } from '../../api/queryKeys';
 import { LibraryCardShell } from './LibraryCardShell';
 import { formatBytes, getDayjsLocale, type StorageDay } from './libraryShared';
