+++ app/ui/src/pages/MigrationCalendar/index.tsx
@@ -18,7 +18,8 @@ import MenuItem from '@mui/material/MenuItem';
 import Button from '@mui/material/Button';
 import { useTranslation } from 'react-i18next';
 import FilterListIcon from '@mui/icons-material/FilterList';
-import { fetchMigrationCalendar, fetchRegionComparison } from '../../api/api';
+import { fetchMigrationCalendar } from '../../api/migrationCalendar';
+import { fetchRegionComparison } from '../../api/weatherRegion';
 import { queryKeys } from '../../api/queryKeys';
 import { SpeciesIcon } from '../../components/SpeciesIcon';
 import { PageHelp } from '../../components/PageHelp';
