+++ app/ui/src/pages/Settings/sections/EBirdSection.tsx
@@ -24,7 +24,7 @@ import type { Settings } from '../../../types';
 import {
   fetchEbirdMappingSuggestions,
   type EbirdMappingSuggestionsResponse,
-} from '../../../api/api';
+} from '../../../api/settingsSession';
 
 type Props = {
   form: ReactFormExtendedApi<Settings, undefined>;
