+++ app/ui/src/pages/SpeciesDirectory/index.tsx
@@ -14,7 +14,8 @@ import Stack from '@mui/material/Stack';
 import TextField from '@mui/material/TextField';
 import Typography from '@mui/material/Typography';
 import { useTranslation } from 'react-i18next';
-import { fetchBirdDirectory, resolveImageUrl } from '../../api/api';
+import { resolveImageUrl } from '../../api/api';
+import { fetchBirdDirectory } from '../../api/speciesOverviewDetections';
 import { queryKeys } from '../../api/queryKeys';
 import { PageHeader } from '../../components/PageHeader';
 import { SpeciesIcon } from '../../components/SpeciesIcon';
