+++ app/ui/src/pages/System/StorageOverview.tsx
@@ -15,8 +15,8 @@ import dayjs, { type Dayjs } from 'dayjs';
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { DatePicker } from '@mui/x-date-pickers/DatePicker';
+import { BASE_API_URL } from '../../api/client';
 import {
-  BASE_API_URL,
   downloadDbBackup,
   purgeStorageRecordings,
   restoreDbBackup,
