+++ app/ui/src/pages/VideoDetails/VideoPlayer/index.tsx
@@ -18,7 +18,8 @@ import Fade from '@mui/material/Fade';
 import Switch from '@mui/material/Switch';
 import FormControlLabel from '@mui/material/FormControlLabel';
 import { Video, VideoSpecies } from '../../../types';
-import { BASE_URL, BASE_API_URL, resolveImageUrl } from '../../../api/api';
+import { resolveImageUrl } from '../../../api/api';
+import { BASE_API_URL, BASE_URL } from '../../../api/client';
 import { ProgressBar } from './ProgressBar';
 import { SpectrogramPlayer } from './SpectrogramPlayer';
 import { useTranslation } from 'react-i18next';
